Create every constructor from scratch
Rather than ripping out everything that appears after the call to the
super-constructor, the constructor is created from scratch. To do so,
the constructor is searched until the appropriate super-constructor is
found. At that point the for each argument type the appropriate neutral
value is pushed to the stack and the super-constructor is invoked.